+++ src/FSharp.Formatting.Literate/Document.fs
@@ -15,6 +15,11 @@ type LiterateCodeVisibility =
   | HiddenCode
   | NamedCode of string
 
+type LiterateParagraphOptions = 
+  { /// Specifies a conditional for inclusion of the snippet paragraph
+    Condition : string option
+  }
+
 /// Additional properties of a literate code snippet, embedded in a
 /// `LiterateParagraph.LiterateCode`. The properties specify how should
 /// a snippet be evaluated and formatted.
@@ -40,27 +45,37 @@ type LiterateCodeOptions =
 /// (such as various special commands to embed output of a snippet etc.)
 type LiterateParagraph =
   /// (*** include:foo ***) - Include formatted snippet from other part of the document here 
-  | CodeReference of string
+  | CodeReference of string * LiterateParagraphOptions
 
   /// (*** include-output ***) - Include output from previous snippet
   /// (*** include-output:foo ***) - Include output from a snippet here 
-  | OutputReference of string 
+  | OutputReference of string * LiterateParagraphOptions
 
   /// (*** include-it ***) - Include "it" value from the subsequent snippet here 
   /// (*** include-it:foo ***) - Include "it" value from a snippet here 
-  | ItValueReference of string 
+  | ItValueReference of string * LiterateParagraphOptions
 
   /// (*** include-value:foo ***) - Include the formatting of a specified value here
-  | ValueReference of string 
+  | ValueReference of string * LiterateParagraphOptions
 
   /// Emebdded literate code snippet. Consists of source lines and options
-  | LiterateCode of Line list * LiterateCodeOptions
+  | LiterateCode of Line list * LiterateCodeOptions * LiterateParagraphOptions
 
   /// Ordinary formatted code snippet in non-F# language (tagged with language code)
-  | LanguageTaggedCode of string * string
+  | LanguageTaggedCode of string * string * LiterateParagraphOptions
 
   /// Block simply emitted without any formatting equivalent to <pre> tag in html
-  | RawBlock of Line list
+  | RawBlock of Line list * LiterateParagraphOptions
+
+  member x.ParagraphOptions =
+    match x with
+    | CodeReference(_,popts) -> popts
+    | OutputReference(_,popts) -> popts
+    | ItValueReference(_,popts) -> popts
+    | ValueReference(_,popts) -> popts
+    | LiterateCode(_,_,popts) -> popts
+    | LanguageTaggedCode(_,_,popts) -> popts
+    | RawBlock(_,popts) -> popts
 
   interface MarkdownEmbedParagraphs with
     member x.Render() = 
@@ -74,11 +89,11 @@ type LiterateSource =
 
 /// Representation of a literate document - the representation of Paragraphs
 /// uses an F# discriminated union type and so is best used from F#.
-type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, errors) =
+type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, diagnostics) =
   // Get the content of the document as a structurally comparable tuple
   let asTuple (doc:LiterateDocument) = 
     List.ofSeq doc.DefinedLinks.Keys, List.ofSeq doc.DefinedLinks.Values,
-    List.ofSeq doc.Errors, doc.Paragraphs
+    List.ofSeq doc.Diagnostics, doc.Paragraphs
 
   /// Returns a list of paragraphs in the document
   member x.Paragraphs : MarkdownParagraphs = paragraphs
@@ -87,7 +102,7 @@ type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, erro
   member x.DefinedLinks : IDictionary<string, string * option<string>> = links
 
   /// Errors
-  member x.Errors : seq<SourceError> = errors
+  member x.Diagnostics : SourceError[] = diagnostics
 
   /// Original document source code
   member x.Source : LiterateSource = source
@@ -109,11 +124,11 @@ type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, erro
   override x.GetHashCode() = (asTuple x).GetHashCode()
 
   /// Clone the document and change some of its properties
-  member x.With(?paragraphs, ?formattedTips, ?definedLinks, ?source, ?sourceFile, ?errors) =
+  member x.With(?paragraphs, ?formattedTips, ?definedLinks, ?source, ?sourceFile, ?diagnostics) =
     LiterateDocument
       ( defaultArg paragraphs x.Paragraphs, defaultArg formattedTips x.FormattedTips,
         defaultArg definedLinks x.DefinedLinks, defaultArg source x.Source, 
-        defaultArg sourceFile x.SourceFile, defaultArg errors x.Errors )
+        defaultArg sourceFile x.SourceFile, defaultArg diagnostics x.Diagnostics )
 
 // --------------------------------------------------------------------------------------
 // Pattern matching helpers
@@ -124,3 +139,4 @@ type LiterateDocument(paragraphs, formattedTips, links, source, sourceFile, erro
 module MarkdownPatterns =
   let (|LiterateParagraph|_|) = function
     | EmbedParagraphs(:? LiterateParagraph as lp, _) -> Some lp | _ -> None
+