+++ src/FSharp.Formatting.Literate/ParseScript.fs
@@ -24,15 +24,15 @@ module internal CodeBlockUtils =
   /// lines & we want to remove all blanks before returning BlockSnippet)
   let private trimBlanksAndReverse lines = 
     lines 
-    |> Seq.skipWhile (function Line[] -> true | _ -> false)
+    |> Seq.skipWhile (function Line(_, []) -> true | _ -> false)
     |> List.ofSeq |> List.rev
-    |> Seq.skipWhile (function Line[] -> true | _ -> false)
+    |> Seq.skipWhile (function Line(_, []) -> true | _ -> false)
     |> List.ofSeq
 
   /// Succeeds when a line (list of tokens) contains only Comment 
   /// tokens and returns the text from the comment as a string
   /// (Comment may also be followed by Whitespace that is skipped)
-  let private (|ConcatenatedComments|_|) (Line tokens) =
+  let private (|ConcatenatedComments|_|) (Line (_, tokens)) =
     let rec readComments inWhite acc = function
       | TokenSpan.Token(TokenKind.Comment, text, _)::tokens when not inWhite-> 
           readComments false (text::acc) tokens
@@ -69,7 +69,7 @@ module internal CodeBlockUtils =
         yield BlockComment (comment.Substring(0, cend))
         yield! collectSnippet [] lines
 
-    | (Line[TokenSpan.Token(TokenKind.Comment, String.StartsWith "(**" text, _)])::lines ->
+    | (Line(_, [TokenSpan.Token(TokenKind.Comment, String.StartsWith "(**" text, _)]))::lines ->
         // Another block of Markdown comment starting... 
         // Yield the previous snippet block and continue parsing more comments
         let cend = findCommentEnd comment
@@ -101,7 +101,7 @@ module internal CodeBlockUtils =
           yield BlockCommand cmds
           yield! collectSnippet [] lines
 
-      | (Line[TokenSpan.Token(TokenKind.Comment, String.StartsWith "(**" text, _)])::lines ->
+      | (Line (_, [TokenSpan.Token(TokenKind.Comment, String.StartsWith "(**" text, _)]))::lines ->
           // Found a comment - yield snippet & switch to parsing comment state
           // (Also trim leading spaces to support e.g.: `(** ## Hello **)`)
           if acc <> [] then yield blockSnippet acc
@@ -111,8 +111,9 @@ module internal CodeBlockUtils =
       | [] -> yield blockSnippet acc }
 
   /// Parse F# script file into a sequence of snippets, comments and commands
-  let parseScriptFile = collectSnippet []
+  let parseScriptFile lines = collectSnippet [] lines
 
+open CodeBlockUtils
 // --------------------------------------------------------------------------------------
 // LiterateScript module
 // --------------------------------------------------------------------------------------
@@ -120,59 +121,70 @@ module internal CodeBlockUtils =
 /// Turns the content of `fsx` file into `LiterateDocument` that contains
 /// formatted F# snippets and parsed Markdown document. Handles commands such
 /// as `hide`, `define` and `include`.
-module internal ParseScript = 
-  open CodeBlockUtils
+type internal ParseScript(parseOptions, ctx:CompilerContext) = 
 
   let getVisibility cmds =
     match cmds with 
-    | Command "hide" _ -> LiterateCodeVisibility.HiddenCode
+    | Command "hide" _ -> LiterateCodeVisibility.HiddenCode 
     | Command "define" name -> LiterateCodeVisibility.NamedCode name
     | _ -> LiterateCodeVisibility.VisibleCode
 
   let getEvaluate noEval (cmds: IDictionary<_,_>)  =
     not (noEval || cmds.ContainsKey("do-not-eval"))
 
+  let getParaOptions cmds =
+    match cmds with 
+    | Command "condition" name when not (System.String.IsNullOrWhiteSpace name)  -> { Condition = Some name }
+    | _ -> { Condition = None }
+
   /// Transform list of code blocks (snippet/comment/command)
   /// into a formatted Markdown document, with link definitions
-  let rec private transformBlocks prevCodeId count noEval acc defs blocks = 
+  let rec transformBlocks prevCodeId count noEval acc defs blocks = 
     match blocks with
     // Disable evaluation for the rest of the file
     | BlockCommand(Command "do-not-eval-file" _)::blocks ->
         transformBlocks None count true acc defs blocks
     
     // Reference to code snippet defined later
-    | BlockCommand(Command "include" ref)::blocks -> 
-        let p = EmbedParagraphs(CodeReference(ref), None)
+    | BlockCommand((Command "include" ref) as cmds)::blocks ->
+        let popts = getParaOptions cmds
+        let p = EmbedParagraphs(CodeReference(ref, popts), None)
         transformBlocks None count noEval (p::acc) defs blocks
 
     // Include output of previous block
-    | BlockCommand(Command "include-output" "")::blocks when prevCodeId.IsSome -> 
-        let p1 = EmbedParagraphs(OutputReference(prevCodeId.Value), None)
+    | BlockCommand(Command "include-output" "" as cmds)::blocks when prevCodeId.IsSome -> 
+        let popts = getParaOptions cmds
+        let p1 = EmbedParagraphs(OutputReference(prevCodeId.Value, popts), None)
         transformBlocks prevCodeId count noEval (p1::acc) defs blocks
 
     // Include output of a named block
-    | BlockCommand(Command "include-output" ref)::blocks -> 
-        let p = EmbedParagraphs(OutputReference(ref), None)
+    | BlockCommand(Command "include-output" ref as cmds)::blocks -> 
+        let popts = getParaOptions cmds
+        let p = EmbedParagraphs(OutputReference(ref, popts), None)
         transformBlocks prevCodeId count noEval (p::acc) defs blocks
 
     // Include formatted 'it' of previous block
     | BlockCommand((Command "include-it" "") as cmds)::blocks when prevCodeId.IsSome -> 
-        let p1 = EmbedParagraphs(ItValueReference(prevCodeId.Value), None)
+        let popts = getParaOptions cmds
+        let p1 = EmbedParagraphs(ItValueReference(prevCodeId.Value, popts), None)
         transformBlocks prevCodeId count noEval (p1::acc) defs blocks
 
     // Include formatted 'it' of a named block
-    | BlockCommand(Command "include-it" ref)::blocks -> 
-        let p = EmbedParagraphs(ItValueReference(ref), None)
+    | BlockCommand(Command "include-it" ref as cmds)::blocks -> 
+        let popts = getParaOptions cmds
+        let p = EmbedParagraphs(ItValueReference(ref, popts), None)
         transformBlocks None count noEval (p::acc) defs blocks
 
     // Include formatted named value 
-    | BlockCommand(Command "include-value" ref)::blocks -> 
-        let p = EmbedParagraphs(ValueReference(ref), None)
+    | BlockCommand(Command "include-value" ref as cmds)::blocks -> 
+        let popts = getParaOptions cmds
+        let p = EmbedParagraphs(ValueReference(ref, popts), None)
         transformBlocks None count noEval (p::acc) defs blocks
 
-    // Include raw code without evaluation
-    | BlockCommand(Command "raw" _) ::BlockSnippet(snip):: blocks -> 
-        let p = EmbedParagraphs(RawBlock(snip), None)
+    // Include code without evaluation
+    | BlockCommand(Command "raw" _ as cmds) ::BlockSnippet(snip):: blocks -> 
+        let popts = getParaOptions cmds
+        let p = EmbedParagraphs(RawBlock(snip, popts), None)
         transformBlocks None count noEval (p::acc) defs blocks
 
     // Parse commands in [foo=bar,zoo], followed by a source code snippet
@@ -187,10 +199,12 @@ module internal ParseScript =
           | _ -> incr count; "cell" + string count.Value
         let opts = 
           { Evaluate = getEvaluate noEval cmds
-            ExecutionCount = None;
+            ExecutionCount = None
             OutputName = outputName
             Visibility = getVisibility cmds }
-        let code = EmbedParagraphs(LiterateCode(snip, opts), None)
+        let popts = getParaOptions cmds
+
+        let code = EmbedParagraphs(LiterateCode(snip, opts, popts), None)
         transformBlocks (Some outputName) count noEval (code::acc) defs blocks
 
     // Unknown command
@@ -205,12 +219,13 @@ module internal ParseScript =
     | BlockSnippet(snip)::blocks ->
         let id = incr count; "cell" + string count.Value
         let opts = { Evaluate=not noEval; ExecutionCount=None; OutputName=id; Visibility=LiterateCodeVisibility.VisibleCode }
-        let p = EmbedParagraphs(LiterateCode(snip, opts), None)
+        let popts = { Condition = None }
+        let p = EmbedParagraphs(LiterateCode(snip, opts, popts), None)
         transformBlocks (Some id) count noEval (p::acc) defs blocks
 
     // Markdown documentation block  
     | BlockComment(text)::blocks ->
-        let doc = Markdown.Parse(text)
+        let doc = Markdown.Parse(text, ?parseOptions=parseOptions)
         let defs = doc.DefinedLinks::defs
         let acc = (List.rev doc.Paragraphs) @ acc
         transformBlocks None count noEval acc defs blocks
@@ -223,11 +238,12 @@ module internal ParseScript =
 
   /// Parse script file with specified name and content
   /// and return `LiterateDocument` with the content
-  let parseScriptFile file content (ctx:CompilerContext) =
-    let sourceSnippets, errors = 
-      ctx.FormatAgent.ParseSource
-        (file, content, ?options = ctx.CompilerOptions, ?defines = ctx.DefinedSymbols)
-    
+  member _.ParseScriptFile filePath content =
+    let defines = match ctx.ConditionalDefines with [] -> None | l -> Some (String.concat "," l)
+    let sourceSnippets, diagnostics = ctx.FormatAgent.ParseSource(filePath, content, ?options=ctx.CompilerOptions, ?defines=defines)
+
+    for (SourceError((l0,c0),(l1,c1),kind,msg)) in diagnostics do
+        printfn "   %s: %s(%d,%d)-(%d,%d) %s" filePath (if kind = ErrorKind.Error then "error" else "warning") l0 c0 l1 c1 msg
     let parsedBlocks = 
       [ for Snippet(name, lines) in sourceSnippets do
           if name <> null then 
@@ -235,4 +251,4 @@ module internal ParseScript =
           yield! parseScriptFile(lines) ]
 
     let paragraphs, defs = transformBlocks None (ref 0) false [] [] (List.ofSeq parsedBlocks)
-    LiterateDocument(paragraphs, "", defs, LiterateSource.Script sourceSnippets, file, errors)
+    LiterateDocument(paragraphs, "", defs, LiterateSource.Script sourceSnippets, filePath, diagnostics)